# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sense API
  description: |
    REST API для социальной сети Sense - платформы для творчества и обмена знаниями.
    
    API поддерживает:
    - Аутентификацию пользователей через JWT токены
    - Создание и управление публикациями (посты, статьи, цитаты)
    - Систему комментариев и лайков
    - Персонализированную ленту с рекомендациями на основе AI
    - Поиск контента и пользователей
    - Загрузку и управление медиа-файлами
    - Профили пользователей с различными ролями
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
...
paths:
...
  /publication/create:
    post:
      tags: [Publications]
      summary: Создать публикацию
      description: Создание новой публикации (пост, статья или цитата)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePublicationRequest'
            example:
              type: "post"
              content: "Сегодня был прекрасный закат в Минске!"
              visibility: "public"
              media_ids: ["123e4567-e89b-12d3-a456-426614174000"]
      responses:
        '201':
          description: Публикация успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /publication/{id}:
    get:
      tags: [Publications]
      summary: Получить публикацию
      description: |
        Получение информации о публикации по ID.
        Аутентификация опциональна - если пользователь авторизован, будет включен статус лайков.
      security: []
      parameters:
        - $ref: '#/components/parameters/PublicationId'
      responses:
        '200':
          description: Информация о публикации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicationResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags: [Publications]
      summary: Редактировать публикацию
      description: Обновление существующей публикации (только автор)
      parameters:
        - $ref: '#/components/parameters/PublicationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePublicationRequest'
            example:
              content: "Обновленный текст публикации"
              visibility: "public"
      responses:
        '200':
          description: Публикация успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Publications]
      summary: Удалить публикацию
      description: Удаление публикации (только автор)
      parameters:
        - $ref: '#/components/parameters/PublicationId'
      responses:
        '204':
          description: Публикация успешно удалена
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /publication/{id}/like:
    post:
      tags: [Publications]
      summary: Поставить лайк
      description: Поставить или убрать лайк с публикации
      parameters:
        - $ref: '#/components/parameters/PublicationId'
      responses:
        '200':
          description: Лайк успешно поставлен/убран
          content:
            application/json:
              schema:
                type: object
                properties:
                  liked:
                    type: boolean
                    description: Лайкнул ли пользователь публикацию
                    example: true
                  likes_count:
                    type: integer
                    description: Общее количество лайков
                    example: 26
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /publication/{id}/likes:
    get:
      tags: [Publications]
      summary: Получить список лайкнувших
      description: Получение списка пользователей, лайкнувших публикацию
      parameters:
        - $ref: '#/components/parameters/PublicationId'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationOffset'
      responses:
        '200':
          description: Список пользователей, лайкнувших публикацию
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  total:
                    type: integer
                    example: 25
                  limit:
                    type: integer
                    example: 20
                  offset:
                    type: integer
                    example: 0
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /publication/{id}/save:
    post:
      tags: [Publications]
      summary: Сохранить публикацию
      description: Добавить публикацию в сохраненные
      parameters:
        - $ref: '#/components/parameters/PublicationId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: string
                  maxLength: 500
                  description: Заметка к сохраненной публикации
                  example: "Интересная статья о фотографии"
      responses:
        '201':
          description: Публикация добавлена в сохраненные
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Публикация добавлена в сохраненные"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Publications]
      summary: Убрать из сохраненных
      description: Удалить публикацию из сохраненных
      parameters:
        - $ref: '#/components/parameters/PublicationId'
      responses:
        '204':
          description: Публикация удалена из сохраненных
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # Комментарии
  /publication/{id}/comments:
    get:
      tags: [Comments]
      summary: Получить комментарии публикации
      description: Получение списка комментариев к публикации с пагинацией
      parameters:
        - $ref: '#/components/parameters/PublicationId'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationOffset'
      responses:
        '200':
          description: Список комментариев
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CommentResponse'
                  total:
                    type: integer
                    example: 15
                  limit:
                    type: integer
                    example: 20
                  offset:
                    type: integer
                    example: 0
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Comments]
      summary: Создать комментарий
      description: Добавить комментарий к публикации
      parameters:
        - $ref: '#/components/parameters/PublicationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
            example:
              text: "Отличная фотография!"
      responses:
        '201':
          description: Комментарий успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен для аутентификации. Получить токен можно через /auth/login или /auth/register
  schemas:
...
    UserRole:
      type: string
      enum: [reader, user, creator, expert, super]
      description: Роль пользователя в системе
      example: "user"
    PublicationType:
      type: string
      enum: [quote, post, article]
      description: Тип публикации
      example: "post"
    VisibilityType:
      type: string
      enum: [public, community, private]
      description: Уровень видимости публикации
      example: "public"

    # Основные модели данных
    User:
      type: object
      required: [id, username, role, registered_at]
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор пользователя
          example: "123e4567-e89b-12d3-a456-426614174000"
        username:
          type: string
          description: Имя пользователя (уникальное)
          example: "john_doe"
        email:
          type: string
          format: email
          description: Email адрес пользователя
          example: "john@example.com"
        phone:
          type: string
          description: Номер телефона пользователя
          example: "+375291234567"
        icon_url:
          type: string
          format: uri
          description: URL аватара пользователя
          example: "https://cdn.sense.social/avatars/user123.jpg"
        description:
          type: string
          description: Описание/биография пользователя
          example: "Фотограф и дизайнер из Минска"
        role:
          $ref: '#/components/schemas/UserRole'
        registered_at:
          type: string
          format: date-time
          description: Дата регистрации
          example: "2024-01-15T10:30:00Z"
        statistic:
          type: object
          description: Статистика пользователя
          properties:
            publications_count:
              type: integer
              minimum: 0
              example: 42
            followers_count:
              type: integer
              minimum: 0
              example: 150
            following_count:
              type: integer
              minimum: 0
              example: 89
            likes_received:
              type: integer
              minimum: 0
              example: 1250
    Publication:
      type: object
      required: [id, author_id, type, publication_date, visibility]
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор публикации
          example: "123e4567-e89b-12d3-a456-426614174000"
        author_id:
          type: string
          format: uuid
          description: ID автора публикации
          example: "123e4567-e89b-12d3-a456-426614174000"
        type:
          $ref: '#/components/schemas/PublicationType'
        content:
          type: string
          description: Основной контент публикации
          example: "Сегодня был прекрасный закат в Минске!"
        source:
          type: string
          description: Источник для цитат
          example: "Альберт Эйнштейн"
        publication_date:
          type: string
          format: date-time
          description: Дата публикации
          example: "2024-01-15T14:30:00Z"
        visibility:
          $ref: '#/components/schemas/VisibilityType'
        likes_count:
          type: integer
          minimum: 0
          description: Количество лайков
          example: 25
        comments_count:
          type: integer
          minimum: 0
          description: Количество комментариев
          example: 8
        saved_count:
          type: integer
          minimum: 0
          description: Количество сохранений
          example: 12
    Comment:
      type: object
      required: [id, publication_id, author_id, text, created_at]
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор комментария
          example: "123e4567-e89b-12d3-a456-426614174000"
        publication_id:
          type: string
          format: uuid
          description: ID публикации
          example: "123e4567-e89b-12d3-a456-426614174000"
        parent_id:
          type: string
          format: uuid
          nullable: true
          description: ID родительского комментария (для ответов)
          example: "123e4567-e89b-12d3-a456-426614174000"
        author_id:
          type: string
          format: uuid
          description: ID автора комментария
          example: "123e4567-e89b-12d3-a456-426614174000"
        text:
          type: string
          description: Текст комментария
          example: "Отличная фотография!"
        created_at:
          type: string
          format: date-time
          description: Дата создания комментария
          example: "2024-01-15T15:30:00Z"
        likes_count:
          type: integer
          minimum: 0
          description: Количество лайков комментария
          example: 3
    MediaAsset:
      type: object
      required: [id, owner_id, mime, created_at]
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор медиа-файла
          example: "123e4567-e89b-12d3-a456-426614174000"
        owner_id:
          type: string
          format: uuid
          description: ID владельца медиа-файла
          example: "123e4567-e89b-12d3-a456-426614174000"
        filename:
          type: string
          nullable: true
          description: Оригинальное имя файла
          example: "sunset.jpg"
        mime:
          type: string
          description: MIME тип файла
          example: "image/jpeg"
        width:
          type: integer
          minimum: 0
          nullable: true
          description: Ширина изображения в пикселях
          example: 1920
        height:
          type: integer
          minimum: 0
          nullable: true
          description: Высота изображения в пикселях
          example: 1080
        exif:
          type: object
          nullable: true
          description: EXIF данные изображения
          example:
            camera: "Canon EOS R5"
            lens: "RF 24-70mm f/2.8L IS USM"
            iso: 400
            aperture: "f/2.8"
            shutter_speed: "1/125"
        created_at:
          type: string
          format: date-time
          description: Дата загрузки файла
          example: "2024-01-15T10:30:00Z"
...
    CreatePublicationRequest:
      type: object
      required: [type, content, visibility]
      properties:
        type:
          $ref: '#/components/schemas/PublicationType'
        content:
          type: string
          maxLength: 10000
          description: Основной контент публикации
          example: "Сегодня был прекрасный закат в Минске!"
        source:
          type: string
          maxLength: 200
          nullable: true
          description: Источник для цитат
          example: "Альберт Эйнштейн"
        visibility:
          $ref: '#/components/schemas/VisibilityType'
        media_ids:
          type: array
          items:
            type: string
            format: uuid
          description: Список ID медиа-файлов для прикрепления
          example: ["123e4567-e89b-12d3-a456-426614174000"]
    UpdatePublicationRequest:
      type: object
      properties:
        content:
          type: string
          maxLength: 10000
          description: Основной контент публикации
          example: "Обновленный текст публикации"
        source:
          type: string
          maxLength: 200
          nullable: true
          description: Источник для цитат
          example: "Альберт Эйнштейн"
        visibility:
          $ref: '#/components/schemas/VisibilityType'
        media_ids:
          type: array
          items:
            type: string
            format: uuid
          description: Список ID медиа-файлов для прикрепления
          example: ["123e4567-e89b-12d3-a456-426614174000"]
    PublicationResponse:
      allOf:
        - $ref: '#/components/schemas/Publication'
        - type: object
          properties:
            author:
              $ref: '#/components/schemas/User'
            media:
              type: array
              items:
                $ref: '#/components/schemas/MediaAsset'
            is_liked:
              type: boolean
              description: Лайкнул ли текущий пользователь эту публикацию
              example: true
            is_saved:
              type: boolean
              description: Сохранил ли текущий пользователь эту публикацию
              example: false
    CreateCommentRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          maxLength: 2000
          description: Текст комментария
          example: "Отличная фотография!"
        parent_id:
          type: string
          format: uuid
          nullable: true
          description: ID родительского комментария (для ответов)
          example: "123e4567-e89b-12d3-a456-426614174000"
    CommentResponse:
      allOf:
        - $ref: '#/components/schemas/Comment'
        - type: object
          properties:
            author:
              $ref: '#/components/schemas/User'
            is_liked:
              type: boolean
              description: Лайкнул ли текущий пользователь этот комментарий
              example: false
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Код ошибки
          example: "validation_error"
        message:
          type: string
          description: Сообщение об ошибке
          example: "Неверные данные в запросе"
        details:
          type: string
          nullable: true
          description: Детали ошибки
          example: "Поле 'email' должно быть валидным email адресом"
  parameters:
    PaginationLimit:
      name: limit
      in: query
      description: Количество элементов на странице
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
    PaginationOffset:
      name: offset
      in: query
      description: Смещение для пагинации
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      example: 0
    PublicationId:
      name: id
      in: path
      description: Уникальный идентификатор публикации
      required: true
      schema:
        type: string
        format: uuid
      example: "123e4567-e89b-12d3-a456-426614174000"
...
  responses:
    BadRequest:
      description: Неверный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "validation_error"
            message: "Неверные данные в запросе"
            details: "Поле 'email' должно быть валидным email адресом"
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "unauthorized"
            message: "Требуется аутентификация"
            details: "JWT токен отсутствует или недействителен"
    Forbidden:
      description: Доступ запрещен
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "forbidden"
            message: "Недостаточно прав для выполнения операции"
            details: "Вы можете редактировать только свои публикации"
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "not_found"
            message: "Ресурс не найден"
            details: "Публикация с указанным ID не существует"
...
security:
  - bearerAuth: []